GRPC/requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
grpcio-reflection==1.54.2